+++ src/lib/InstantSearch.js
@@ -203,7 +203,18 @@ Usage: instantsearch({
   }
 
   /**
-   * The start methods ends the initialization of InstantSearch.js and triggers the
+   * The refresh method clears the cached answers from Algolia and triggers a new search.
+   *
+   * @return {undefined} Does not return anything
+   */
+  refresh() {
+    if (this.helper) {
+      this.helper.clearCache().search();
+    }
+  }
+
+  /**
+   * The start method ends the initialization of InstantSearch.js and triggers the
    * first search. This method should be called after all widgets have been added
    * to the instance of InstantSearch.js. InstantSearch.js also supports adding and removing
    * widgets after the start as an **EXPERIMENTAL** feature.
@@ -274,6 +285,9 @@ Usage: instantsearch({
     this.helper = helper;
     this._init(helper.state, this.helper);
     this.helper.on('result', this._render.bind(this, this.helper));
+    this.helper.on('error', e => {
+      this.emit('error', e);
+    });
 
     this._searchStalledTimer = null;
     this._isSearchStalled = true;